requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.19.0
